Seperate agentic and classic notebook into two components (#159)
