+++ common/types/notebooks.ts
@@ -178,3 +178,7 @@ export interface PERAgentInvestigationResponse {
   hypothesis: PERAgentHypothesisItem;
   operation: 'CREATE' | 'REPLACE';
 }
+
+export interface NotebookComponentProps {
+  showPageHeader?: boolean;
+}
