+++ public/components/notebooks/components/__tests__/classic_notebook.test.tsx
@@ -20,7 +20,7 @@ import {
   notebookPutResponse,
   runCodeBlockResponse,
 } from '../../../../../test/notebooks_constants';
-import { Notebook, NotebookProps } from '../notebook';
+import { ClassicNotebook, ClassicNotebookProps } from '../classic_notebook';
 import {
   applicationServiceMock,
   chromeServiceMock,
@@ -139,7 +139,7 @@ jest.mock('../../../../../public/services', () => ({
   })),
 }));
 
-const ContextAwareNotebook = (props: NotebookProps & { dataSourceEnabled?: boolean }) => {
+const ContextAwareNotebook = (props: ClassicNotebookProps & { dataSourceEnabled?: boolean }) => {
   const applicationMock = applicationServiceMock.createStartContract();
   return (
     <OpenSearchDashboardsContextProvider
@@ -188,7 +188,7 @@ const ContextAwareNotebook = (props: NotebookProps & { dataSourceEnabled?: boole
         },
       }}
     >
-      <Notebook {...props} />
+      <ClassicNotebook {...props} />
     </OpenSearchDashboardsContextProvider>
   );
 };
@@ -201,7 +201,7 @@ describe('<Notebook /> spec', () => {
   const history = jest.fn() as any;
   history.replace = jest.fn();
   history.push = jest.fn();
-  const defaultProps: NotebookProps = {
+  const defaultProps: ClassicNotebookProps = {
     openedNoteId: '458e1320-3f05-11ef-bd29-e58626f102c0',
     showPageHeader: true,
   };
